+++ src/sapwebguimcp/fmcatalog/search.py
@@ -23,7 +23,7 @@ class FMSearchResult:
     """A single search result with relevance score."""
 
     fm: FunctionModuleEntry
-    score: int
+    score: float
     match_reason: str
 
 
@@ -96,7 +96,7 @@ def search_function_modules(
         if score == 0 and fm.description:
             fuzzy_score = fuzz.WRatio(query, fm.description, score_cutoff=50)
             if fuzzy_score:
-                score = 5 + int((fuzzy_score - 50) * 9.0 / 50.0)
+                score = 5.0 + (fuzzy_score - 50) * 9.0 / 50.0
                 match_reason = "fuzzy description"
 
         if score > 0:
